Address follow-up trio review
